+++ src/Digdir.Domain.Dialogporten.GraphQL/ServiceCollectionExtensions.cs
@@ -2,9 +2,7 @@ using AppAny.HotChocolate.FluentValidation;
 using Digdir.Domain.Dialogporten.GraphQL.EndUser;
 using Digdir.Domain.Dialogporten.GraphQL.EndUser.DialogById;
 using Digdir.Domain.Dialogporten.GraphQL.EndUser.MutationTypes;
-using Digdir.Domain.Dialogporten.GraphQL.EndUser.SearchDialogs;
 using Digdir.Domain.Dialogporten.Infrastructure.Persistence;
-using HotChocolate.Execution.Configuration;
 
 namespace Digdir.Domain.Dialogporten.GraphQL;
 
@@ -25,26 +23,4 @@ public static class ServiceCollectionExtensions
         .AddInstrumentation()
         .InitializeOnStartup()
         .Services;
-}
-
-internal static class IRequestExecutorBuilderExtensions
-{
-    internal static IRequestExecutorBuilder AddErrorTypes(this IRequestExecutorBuilder builder)
-    {
-        var errorInterfaces = new[]
-        {
-            typeof(IBulkSetSystemLabelError),
-            typeof(ISetSystemLabelError),
-            typeof(ISearchDialogError),
-            typeof(IDialogByIdError)
-        };
-
-        var errorTypes = typeof(ServiceCollectionExtensions).Assembly
-            .GetTypes()
-            .Where(t => !t.IsAbstract && errorInterfaces
-                .Any(i => i.IsAssignableFrom(t)) && t.IsClass)
-            .ToList();
-
-        return errorTypes.Aggregate(builder, (current, type) => current.AddType(type));
-    }
 }
